@@ -15,7 +15,7 @@
         }
     ],
     "require": {
-        "php": ">=5.5.0",
+        "php": ">=7.0.0",
         "egulias/email-validator": "~2.0"
     },
     "require-dev": {
