@@ -4,7 +4,7 @@ require_once 'swift_required.php';
 
 //This is more of a "cross your fingers and hope it works" test!
 
-class Swift_DependencyContainerAcceptanceTest extends \PHPUnit_Framework_TestCase
+class Swift_DependencyContainerAcceptanceTest extends PHPUnit\Framework\TestCase
 {
     public function testNoLookupsFail()
     {
