@@ -1,6 +1,6 @@
 <?php
 
-class Swift_Events_TransportChangeEventTest extends \PHPUnit_Framework_TestCase
+class Swift_Events_TransportChangeEventTest extends \PHPUnit\Framework\TestCase
 {
     public function testGetTransportReturnsTransport()
     {
