@@ -1,6 +1,6 @@
 <?php
 
-class Swift_Mime_Headers_DateHeaderTest extends \PHPUnit_Framework_TestCase
+class Swift_Mime_Headers_DateHeaderTest extends \PHPUnit\Framework\TestCase
 {
     /* --
     The following tests refer to RFC 2822, section 3.6.1 and 3.3.
