@@ -1,6 +1,6 @@
 <?php
 
-class Swift_Plugins_PopBeforeSmtpPluginTest extends \PHPUnit_Framework_TestCase
+class Swift_Plugins_PopBeforeSmtpPluginTest extends \PHPUnit\Framework\TestCase
 {
     public function testPluginConnectsToPop3HostBeforeTransportStarts()
     {
