@@ -1,6 +1,6 @@
 <?php
 
-class Swift_Plugins_Reporters_HtmlReporterTest extends \PHPUnit_Framework_TestCase
+class Swift_Plugins_Reporters_HtmlReporterTest extends \PHPUnit\Framework\TestCase
 {
     private $html;
     private $message;
